mobile_store_app_windows/requirements.txt
streamlit>=1.32.0
supabase>=2.3.0
pandas>=2.0.0
altair>=5.0.0
python-dotenv>=1.0.0
zxing-cpp>=2.0.0
pyzbar>=0.1.9
Pillow>=10.0.0
opencv-python>=4.8.0
streamlit-webrtc>=0.47.0
